errors.py file added
